Wallet Async Fixes
The websockets library uses separate functions for sending and receiving. This means that between a thread sending a request and receiving the response, another thread's response may come first. This results in incorrect responses being received. To prevent this, I made a websocket handler that is the only thread that performs sends and receives. A send/receive can be performed in one function call via wsRequest(). Requests are placed in a pool, which the loop iterates through, grouping requests to responses. When the loop inside of wsRequest() detects that the response has been provided, it returns it and removes the initial request.