Fixed Ledger Corruption issues
Now validates the ledger on startup, which meant I discovered some invalid blocks that are buried too deep to be removed, so are now marked as valid.

There is now a locking mechanism meaning that a node will only accept one pending transaction for an address at a time, and will reject/vote against all other transactions. Closes #40