Handle Internal Errors
Handle any exception that happens during request handling and return a message saying there was an internal error.

Also return defaults for some requests when they are performed on a non-existent address. For example, return an empty list if asking for the transactions belonging to a non-existent account instead of erroring.